Fix setup.sql/seed.sql: cast NULL mrp to numeric
All-NULL VALUES column inferred as text, clashing with numeric mrp column.
Garden Green Calm variant insert failed with 42804. Use NULL::numeric.

+++ supabase/seed.sql
@@ -67,8 +67,8 @@ insert into public.product_variants (product_id, weight_grams, label, price, mrp
 select id, w.grams, w.label, w.price, w.mrp, w.sku, w.stock, w.sort
 from p
 cross join (values
-  (250,  '250g', 349, NULL, 'GGC-250', 90, 1),
-  (500,  '500g', 649, NULL, 'GGC-500', 60, 2)
+  (250,  '250g', 349, NULL::numeric, 'GGC-250', 90, 1),
+  (500,  '500g', 649, NULL::numeric, 'GGC-500', 60, 2)
 ) as w(grams, label, price, mrp, sku, stock, sort);
 
 insert into public.product_images (product_id, url, alt, sort_order)
